frontend: store/skill_form: Rename {,_}bond.js to avoid name conflict
